Update Folder Locker using java
The code provided is a Java program called "FolderLocker" that allows you to lock and unlock a folder by creating a hidden, locked version of the folder. It prompts the user to enter the path of the folder they want to lock or unlock, as well as a password for accessing the locked folder.

When the folder is locked, it renames the original folder by appending "_locked" to its name. It also creates a "password.txt" file inside the locked folder, which contains the provided password.

To unlock the folder, the program checks if the entered password matches the one stored in the "password.txt" file. If the passwords match, it renames the locked folder back to its original name.

This code can be useful for securing sensitive folders by preventing unauthorized access. It can be used as a starting point to build a more robust folder locking mechanism or as a standalone tool.

If you intend to use this code in a GitHub repository, you can create a README.md file and include a description of the code, its purpose, and how to use it. Additionally, you can provide instructions on any dependencies or prerequisites required to run the program successfully.