NumberGuessingGame.java
Number Guessing Game in java-
This Java program implements a simple number guessing game. It generates a random number within a predefined range and prompts the user to guess the number. The user has a limited number of attempts to guess correctly. After each guess, the program provides feedback on whether the guess was too high or too low. If the user guesses the correct number within the allotted attempts, their score increases. The game continues until the user either guesses the number correctly or exhausts all attempts.